+++ src/main/java/domain/participant/BettingMoney.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.participant;
 
 public class BettingMoney {
