+++ src/main/java/domain/participant/Decision.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.participant;
 
 import java.util.Arrays;
 import java.util.Objects;
