+++ src/main/java/domain/participant/Hand.java
@@ -1,5 +1,6 @@
-package domain;
+package domain.participant;
 
+import domain.card.Card;
 import java.util.Deque;
 import java.util.LinkedList;
 import java.util.List;
