+++ src/main/java/domain/participant/Participant.java
@@ -1,6 +1,8 @@
-package domain;
+package domain.participant;
 
+import domain.card.Card;
 import java.util.List;
+import java.util.Objects;
 
 public abstract class Participant {
 
@@ -30,5 +32,22 @@ public abstract class Participant {
 
     public String name() {
         return name.value();
+    }
+
+    @Override
+    public boolean equals(final Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+        final Participant that = (Participant) o;
+        return Objects.equals(name, that.name);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(name);
     }
 }
