+++ src/test/java/domain/card/SuitTest.java
@@ -1,7 +1,8 @@
-package domain;
+package domain.card;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import domain.card.Suit;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
