+++ src/test/java/domain/participant/PlayerTest.java
@@ -1,7 +1,10 @@
-package domain;
+package domain.participant;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import domain.card.Card;
+import domain.card.Number;
+import domain.card.Suit;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
